Changed a way redis keeps session data
Due to the resource conflict resolution problem it must be possible to
store more than one session data under the same USR key. Because of that
all hashes stored in redis were converted to sets, and session data
must be stored in the DB twice: under US and USR key.