Add stubs for owner test cases
Basically, ejabberd behaves differently than the XEP requires it to.

The room_creation_not_allowed is OK (functionally), but the error message
is different than required by the spec.

The cant_enter_locked room fails, as ejabberd doesn't support locked rooms.
The fact that mod_muc_room is a gen_fsm not a gen_server is a hint that it maybe
was planned, but wasn't actually implemented.

The remaining two test cases (create_instant_room, create_reserved_room)
are unfinished.