Make per-host config Host terms binaries esl/ejabberd#23
Global config hosts are converted to binaries, but not those in host_config terms.
This just fixes the term defining which host it applies to. There might be host terms
in the config terms that should be converted too. Without converting these to binary,
the config lookup for per-host config was 'undefined'.
This also tries to enforce binaries when processing terms to fail early instead of
silently like it did.
Host can also be 'global'