Skip registered_users test for external auth
TL;DR: fixing ejabberd is too much effort for too little gain.

``ejabberd_auth_external`` can optionally cache registration data in mnesia.
However, if it's not configured to do so, the user must assume that some
of its callbacks won't work as they rely on the mnesia cache.

Fixing this reliance would change the API of ``extauth.erl``.
Making the option an obligation to use the cache would incur
memory/performance overhead.

No client is known to use the extauth mechanism as is, they either require
custom authentication modules or use internal/odbc auth.

Given all of the above, it's most convenient to just skip this test when
the server runs with ``ejabberd_auth_external``.

Similar limitation apply to ``ejabberd_auth_anonymous`` - in its case
the test is also skipped.