Update Rebar link...
... to the one documented in the current rebar project's README:
  https://github.com/rebar/rebar/blob/837df640872d6a5d5d75a7308126e2769d7babad/README.md#downloading

The old link has not been updated for a while:
  $ curl -s -I http://cloud.github.com/downloads/basho/rebar/rebar | grep Last-Modified
  Last-Modified: Mon, 23 Apr 2012 23:45:48 GMT

Probably this was the reason why clean clones were not working and a
copy of rebar was checked in in Nov 2012 - see commit 2535a59.  I am
not removing the checked in version of the rebar binary as I have not
tested whether the current codebase works with an up-to-date version
of rebar.