Fix parallel execution of multiple mongooseimctl instances
It seems that jot is commonly available on MacOS X
while flock on Linux.

On MacOS X jot tries to come up with a uniqe nodename by
randomization - this does not guarantee uniqeness,
but has a good chance.

On Linux, without CONNLOCKDIR present flock is not able
to determine a unique nodename for mongooseimctl script instance.
This causes all instances to use the same default nodename
and this causes issues when writing scripts using mongooseimctl.