rename config options and use one riak pool
instead of creating pool of pools - each pool connecting to the one
riak node, we move the responsiblity to the user. The user can use load balancer
and point mongoose towards it. This simplifies reconfiguration of running Mongoose
when the riak node is either added or removed. With this approach MiM has nothing
to do, everything is handled by load balancer.