Change a misleading character literal found in a sensitive place
NOTE: the code fragments herein are enclosed between backticks.

The character literal for SPACE was not clear: `$\ ` especially
because these three characters preceded the arrow in a case clause's
head. So, the source looked something like this:

      `when Char /= $\ ->`

This is replaced with the macro ASCII_SPACE_CHARACTER which expands to
`$\s`, which is clearer.

The procedure in question makes an attempt to parse quoted commandline
arguments, utilising escape sequences like `\"`. In this context it
was a particularly sensitive place to find a literal that can easily
be misread as the literal for the backslash escape character.