ejabberd_c2s: Omit features on required STARTTLS.
According to RFC 6120, section 5.3.1 it's not recommended to send
features before the client has finished STARTTLS:

http://xmpp.org/rfcs/rfc6120.html#tls-rules-mtn

I've stumbled on this with Gajim 0.16.3 on Windows but haven't yet able
to reproduce this with the very same client on NixOS, Ubuntu and Debian.

While this originally was/is an issue with Gajim, MongooseIM really
should implement what is recommended by the RFC, regardless of whether
clients out there break because of this (or see it as an open invitation
to send cleartext passwords).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>