Quote arguments ONLY for the extras that `mongooseimctl` offers over `mongooseim`
The use of "$@" in `mongooseimctl` preserves the double quotes
surrounding arguments containing spaces. This is detrimental when
`mongooseim` is subsequently invoked with these, but fine for
ivokations that are deligated to `ejabberd_ctl`. The procedures
`args_join_string` and `args_join_xml` in `ejabberd_ctl` piece back
arguements delimited with double quoutes (").

Thus we scan arguments into two shell variables, as-is, and
quoted: then pass these to `mongooseim` and `mongooseimctl`
respectively.