Use specific timeouts for odbc queries
Currently the timeout values has been hardcoded and set to 5 seconds.
Setting a timeout value (any value less than infinity) is necessary to
avoid blocking a connection worker forever in the event of underlying
connection disappears without a trace. This might happen if an
intermediate router or load balancer decides to drop the connection.