Improve test cases for user cancelation
Test that Mongoose responds with the failure stanzas given XEP 0077
upon a bad cancelation request. Changing the configuration in the
"not-allowed" test should correspond better with the XEP: ALL users
are denied cancelation, rather than just stopping the the module.

The tests don't fail in quite the right way, so this is [WIP].