Drive Cover completely from run_common_test, not from within test suites
Calling mongoose_cover_helper:analyze/0 causes Cover to unload cover
compiled modules.
This in turn leads to badfun errors and inconsistencies in code state between nodes,
if code is unloaded on one node, but not on the other(s).

In other words, Cover should be enabled and disabled on
__all nodes which make up the cluster__ at the same time.
The emphasized part is important: a federated node can still have Cover
turned on / off independently as funs aren't shared through XMPP federation.
This is the case in s2s_SUITE.