Add privacy unavailable msgs for 2nd order effects
XEP-0016 2.11 says that contacts who are newly blocked from receiving
presence information should receive an 'unavailable' presence message at
the time of blocking. The XEP is not explict about this including
cases where privacy lists have not changed, but a user's subscription or
group membership have, making them newly subject to blocking. However it
seems like a reasonable case to cover, given that presumably the idea
behind the requirement is that a user should not be left thinking a user
who has blocked them is indefinitely available.