Write a more complex test scenario and change implementation
The more complex scenario has three participants and mixes the two
protocols (i.e. it demonstarates the interaction between an HTTP
administrator and XMPP clients). The implementation has changed to
build and send stanzas because the logic for MUC is intertwined and
not exposable (not without considerable effort). I.e. it's written
only with XMPP and ejabberd internals in mind.