create session item only when we are sure the stream can be resumed
Before the change it was possible to have session leak due to
failure in stream resumption.
Previously the entry to session table was created before checking the
'h' parameter. If this parameter was too high it resulted in process
termination but the entry from session table was not clean as the
process was not in a stat allowing to clean the entry.