Make DB connection pool configuration independent from the host config.
To declare a DB connection pool, use on of the following formats:

  {pool, odbc, PoolName}.
  {pool, odbc, PoolName, Options}.

PoolName is an atom and Options is a list of ODBC options. Options can be still specified at the top level.

1. All hosts use the ‘default’ pool by default.
  To change this, use the {odbc_pool, Option} option (top-level or per-host).

2. This means that setting ‘odbc_server’ is not enough for a DB connection to work. The default pool has to be declared as well.

3. The ODBC options specified per-host are not allowed anymore. They have been replaced with the Options element of the pool declaration.

To update the previous config with ODBC enabled to the new format, add the following line to ejabberd.cfg:

  {pool, odbc, default}.