Accum stage five (#1306)
* Pass packet to be send to routing function along with accumulator
(in case they are different); we get rid of to_send key in nearly all
places. The only use for it now is deep within routing modules, from where
the accumulator does not return, while having outgoing stanza inside it
simplefies coding. Can be change later on, if requested.

* documentation, code samples

* bugfixes

* upgraded sm

* cache privacy check keyed by arguments (needed because e.g. amp runs privacy check multiple times with varying args)

* improved privacy check to allow checking arbitrary stanzas

* cleanup of mongoose_acc module

* acc in routing errors

* removed a number of deprecated things

* cleanup and bugfixes

* switch to run_fold, use of mongoose_acc

* as per code review

* format, link

* bugfix

* [skip ci] doc rev

please see if the meaning is still there;)

* [skip ci] update diagram desc

I just looked at it;P

* minus  the