[skip ci] Delete TODO from gen_mod
What kind of other function would we need here?
We have `get_opt/2`, `get_opt/3` and `get_opt/4` for
fetching options. In case one needs a default value when the one
he was looking for was not found, he should use the latter two.
If he's sure it exists he wants to crash when not found (that would
mean somethings wrong with his predictions about the program).
We could also return an atom `undefined` but - does it make a
difference?