Fix failing mod_mam_muc when only MUC mam enabled
We now check wether the message is of type "groupchat" as we don't
support chat markers for these. Before this commit it wasn't checked
and the module mod_mam_params wasn't even compiled (it is compiled in
mod_mam and we enabled only mod_mam_muc) which led to crash.