Fail in password_can_be_changed which gave false positives
Due to:

- no bob_altpass in Config's escalus_users
- escalus_story:apply_w_arity_check/2 logic:

	apply_w_arity_check(Fun, Args) when is_function(Fun, 1) ->
		case length(Args) of
			1 -> apply(Fun, Args);  %% Fun expects one logged-in user
			_ -> apply(Fun, [Args]) %% Fun expects list of users
		end;
	apply_w_arity_check(Fun, Args) when is_function(Fun) ->
		apply(Fun, Args).

`escalus_story:clients_from_resource_counts/2` did not expand `bob_altpass` to a client spec.
An empty list was returned instead.
Length of an empty list is 0, which made control flow through the _Fun expects list of users_
clause above, which in turn lead to `StoryFun` being passed a parameter for _Bob,
but the value of this parameter was actually an empty list,
not a successfully connected client.

This change makes this problem obvious by failing with a function_clause error.