Introduce 'foreground' mode #1775
* Introduce 'foreground' mode

In rebar the 'foreground' mode is usually used when the erlang app is
running as a daemon within a supervising service such as systemd or
upstart. To work correctly within those services some flags for 'erl'
are necessary so that for example unix pipelines can be used. This
patch introduces the 'foreground' mode as known from rebar which is
very similar to live/console but sets the following additional flags:

    -noshell -noinput +Bd

Those are documented at:

    http://erlang.org/doc/man/erl.html

The very same flags are also used in relx and distillery:

https://github.com/bitwalker/distillery/blob/master/priv/libexec/commands/foreground.sh#L18
https://github.com/erlware/relx/blob/86d415f33c500d432d6c9fa2fcbff3bb2fff3e8f/priv/templates/extended_bin#L623

It should definitely be thought about adjusting the systemd unit file so
that 'foreground' is used instead of 'start'. Otherwise the beam process
is not considered as main PID and when the beam is killed you are not
notified properly via the standard systemd tooling (e.g. 'systemctl
status' returns 0).

* Add some basic documentation for 'foreground' mode

* Mention new 'foreground' mode in list of commands