Proposed fix of the root cause of iq looping issues.
The point is that iq_query_info was cached in the accumulator and
returned the same all the time, which is usually ok, but not after call
to jlib:make_error_reply because it changes stanza type. Therefore on
the second iteration information returned by the call to
mongoose_acc:get(iq_query_info, Acc) in ejabberd_local:process_iq/4 was
not accurate, and we entered the loop.