Fix Travis build / disallow direct entry into the container
This commit unifies the way Centos6 with extra software collections
and Centos7/Debian are started by the `run` script.
It also changes the `run` script interface to immediately run the build,
without dropping the calling user into a build container shell and requiring
to run another script in that env.