Accumulate number of tests which passed due to retries
We need this number as some suites might not use groups at all.
In such a case, we can't simply decide if a build is green or red
just based on group stats, since the groups stats won't tell us truth.
We have to make sure we take into account the number of failed test cases
and subtract from it the number of test cases which were retried
and eventually passed.