Don't raise error when Mnesia dir differs with app env directory (#1904)
This patch removes the part of code which raises an error if Mnesia
directory reported by `mnesia:system_info(directory)` is the same as
`application:get_env(mnesia, dir)` when deleting Mnesia files. From
now on only a warning message is emitted.

The check sometimes resulted in unexpected errors. Users set Mnesia
directory path using application env `mnesia.dir`. This path might be a
relative path (it might be even an atom!). When Mnesia starts, it takes
this value, converts it to absolute path and stored it in initial
config. This is what `mnesia:system_info(directory)` returns, thus the
check failed in simple situations when someone sets `mnesia.dir` to
relative path. The check only succeeded in cases when `mnesia.dir` path
is a normalized, absolute path.