Adjust "conflict_check_failed" log severity
See #1963. There are error logs occuring when a client attempts to
reconnect in a very short timespan and at the same time a stanza is
routed to the old invalid connection.

As this is just an application/client oddity the log severity should be
lowered to "warning". This is simply nothing to alert an operations guy
in the middle of the night.

There are actually a lot more logs like this but this one currently
stands out ;).