Set user 'unavailable' after ping timeout
If a user is timed out due to mod_ping then there is no presence
information broadcasted, the session is simply killed or hibernated
(when stream management is used). This patch lets mod_ping send an
'unavailable' presence instead of the user such that other peers
are informed about the timed out session.