Replace dots in "host" part of metric name with underscores.
Underscores are not a valid part of a domain name, so they
are safe to replace dots to in the metric name (i.e. this can't lead
to a collision with an existing host name).

Replacing dots with underscores prevents hosts from being treated
as multiple separate metrics nodes e.g. by Graphite and Grafana.