Do not rely on tty command which might be broken in containers
`tty` might or might not work depending on whether a Docker container
or a Kubernetes pod is run with `-t` / `tty: true`.
Moreover, it depends on the flags passed initially to `run`
and also to flags passed later to `exec` when attaching
a shell to an already running container.
See https://github.com/moby/moby/issues/8755 for details.

It's more convenient to just generate a random ID like relx does
in scripts it generates for releases.