Fixed the unstable tests.
The issue is due to a race condition, between the muclight routing, and
the availability of the user. When Alice sends a message to the muclight
room, the message is processed and routed within her own c2s process
—unlike regular muc, where a room gets its own c2s process— and if
during the routing process, Kate reconnects, there's no reason for the
message not to be delivered to Kate. So, to ensure that Kate reconnects
_after_ the whole routing is finished, it is as simple as to make Alice
wait for her own Stanza: this is how muclight works, a user's c2s
process routes everything as needed and, at the last step, it routes to
itself.