Extend timeout for the io operation.
When debugging the errors, I discovered that the operation was failing
by returning `{error, terminated}` from `file:read_line(standard_io)`,
which was not matching any clause. Then I realised that this could only
happen by `group_leader` dying, which could only happen by the app being
signaled to die. When debugging, I discovered that `group_leader` was
dying with error `epipe`, which is, that a Unix pipe has been closed on
the other side all the while this side of the pipe still tries to read
from it. What is the other side of the pipe? Common_test!

So in the end, I noticed that successful tests were always passing in
like 1.9s, when the timeout is 2s. So when not receiving anything in
time, the test was failing, the process owning the port was dying, hence
signalling the port with a `sigpipe`, hence telling the OS to end the
pipe, hence signalling the other side with a `epipe` when it was
insisting in continuing, hence, _failing!_.

Ah, so much investigation for just a timeout.