CI stabilisation — ignore known failure
If the test fails with the known reason, simply exit the escalus story
and return `{skip, Reason}`. This way, we don't get CI in red any more
for the known issue, but if it ever fails with a different one, then CI
will get red and we will know about it.