Wait for all room operations
Ah, very important, the RMQ subscription might happen in the middle of
the room stanzas processing and hence the affiliation messages might be
delivered anyway. By waiting for the stanzas for all users and
subscribing to RMQ only after that, THEN we can be sure that RMQ won't
have dirty messages. Tested 10000 times successfully.