Fix unstable test
On `escalus:story`, escalus will check if the users are connected by
`escalus_connection:is_connected/1`, and if this returns true, it will
attempt to receive the stream end stanza. But this is then a race
condition, as this stanza was already received, between escalus_tcp
closing the connection before `escalus:story` reaches this check. Easy
fix, simply wait for closing after receiving the stream end stanza.