Remove a leftover config file
'mongooseim.cfg.token-support' was created to test
token-based authentication (dba4a3d9e).

This file is unused now - it is not referenced anywhere.
Sometimes it was updated with config format, and sometimes not.
As a result, it is also not a valid config file anymore.

There is no reason to keep it any longer.