introduce a module for hooks sanity checks
It's not possible to have a static analysation of hooks execution.
Tests does not always check if a hook was executed and finished with
expected arguments.
I thought it'd be worth to add a module (enabled only on dev nodes)
checking the hook arguments before any other hook is run and after all hooks.

This commit introduces only verification of `user_send_packet` hook.