Remove dead ejabberd_auth:remove_user/3
Currently there is no way to call this function other than from
MongooseIM console.
In the past (long time ago) this function was used to verify if a user
is allowed to remove itself by providing the password.
Nowadays there is no admin interface for that nor XMPP protocol / XEP.