ejabberd_c2s call standard mongoose_transport
`ejabberd_c2s` is calling starttls in two different ways: in one, it
calls it through the mongoose_transport behaviour, which in turns calls
starttls on the module that was given to the first as a parameter, and
on the second it calls it directly over such module, which seems like an
inconsistent way to call such function.

All the modules we have that implement mongoose_transport actually
implement two `starttls` callbacks, and even in some, like in
mod_websockets, we mention in the comments that some callbacks are just
to make them compatible with ejabberd_sockets, but the `starttls/3`
callback is not required in the behaviour definition. So that's a first
code cleaning.