Implement logger_ct_backend
The purpose is to have a way from tests to capture logs in a node, to
verify that certain event was actually logged. This was done in lager,
attaching a handler to the `lager_event` gen_event handler.

Now, we instead create a logger handler, following the necessary
callbacks as described in the logger docs. We're creating a gen_server
that will keep track in its state of the Pids that have requested being
reported of events of a certain loglevel. This gen_server is started in
`adding_handler/1`, which is an optional callback of a logger's handler,
which is call when we actually do `logger:add_handler/3`.

Processes requesting to capture log events send a message to this
gen_server with their pid desired loglevel, and the gen_server keeps
them in its state. When logger calls the `log/3` callback, this is
dispatched async to this gen_server, who in turn dispatches this to all
the processes that have requested him this loglevel.