Rewrite ejabberd_loglevel to mongoose_logs
The API has changed considerably. The first thing to notice is that
loglevels are now syslog (https://tools.ietf.org/html/rfc5424)
compliant, that is, we're ditching the old 5 loglevels and using now
what logger, even earlier lager, and all the logging industry, uses as
severity levels.

Considering that we're now using OTP's logger, we want to design this
helper module in line with logger's design. Logger exposes a primary
config, which is the one we set and get loglevels into; the getter
simply returns the atom with the severity keyword, the setter simply
returns ok if the loglevel given was correct. We don't set nor return
loglevels for all backends, since in logger, all the handlers will
default to the primary config if they didn't have a custom parameter.

If a specific backend (handler) needs to have a custom specification, we
might assume that the operator knows what he's doing and might do so
using logger's API directly. Such functionality was not available with
ejabberd_loglevel anyway.

Setting a loglevel per module is possible only (at least in a simple
way), through this primary config. If any handler is setting a more
specific loglevel, it might override the loglevel set by the module.
That is, when an event is logged, it will be filtered against the global
loglevel, or its local per-module loglevel, if there's one; if this
passes, it will be again filtered per handler, and a handler might as
well ignore the event.

One more function, `loglevel_number_keyword/1`, is exposed only to
convert between a loglevel's keyword and its corresponding number, as
such functionality is basic and not included in logger itself.