Choose config file type by extension
Previously the '.cfg' file extension was stripped
and if there was a '.toml' file with the same root name,
it was used instead.

This was counter-intuitive and now the config file name
is configured explicitly - it is 'mongooseim.toml' by default.
There is no implicit fallback to 'mongooseim.cfg' anymore.
Having one explicitly configured config path is less error prone.