Fix error during config reload
ctl shells open a hidden Erlang node and connect it to the mim node
through erlang distribution, and unless specified with
`-noinput -noshell` flags, it will redirect its log messages to the
connected node as well. When checking the output of the ctl command, we
were checking that the expected return value, `"done"`, was at the
beginning of the returned string. But this is imprecise, as the order is
not defined. Lager was leaving the log values at the end, while logger
seems to be leaving them at the beginning. What we want to is to check
whether the expected return value is part of the actual return,
regardless of the order.