Update all documentation references to logging
Note that some pieces were also reworded or removed altogether. For
example, I replaced all advise to use numbers as loglevels to using
their keyword counterparts, which are more explicit.

All documentation referring to `syslog` is removed, as we don't support
syslog anymore, though we expect to recover this promptly. Then, it will
be recovered, if not rewritten.