Fix a bug in host_config options
Each handler can return any number of config options.
In case of host_config, they are functions which return
  '(local_)config' records for a given host.

So far the code accepted only one function,
which broke the functionality of 'host_policy',
that has multiple records resulting from a single config section.

This was the case for other options as well, the conincidence is
that they were just not configurable per domain.

There is no new functionality here, only a fix of what was incomplete.