Unify does_user_exist API
Up to now, in ejabberd_auth, we've had two almost identical functions,
`is_user_exists` and `does_user_exist`. This is redundant and confusing,
so I've just default back to an overloaded `does_user_exist`, where one
takes a full JID, and the other takes a username and server and creates
a JID and calls the first. The goal would be to eventually get rid of
the non-JID one and use only typed data-structures.