Reduce stringprepping in auth modules
PROBLEM: if we pass forward an “untyped” binary for the user, server,
and/or resource of a jid, we don't know if it has been stringprepped or
not, so we are always stringprepping again and again things that
probably were already, wasting CPU time. This was seen in profiles when
working on the ejabberd_sm case.

SOLUTION: Rewrite function calls and hooks to pass a #jid{} record,
which is clear about its pieces, and extract the stringprepped chunks
only on the place where it is used.