Do not use the 'cfg' format in revproxy_SUITE
This test suite checks the functionality of mod_revproxy, which
requires the 'mod_cowboy' handler for the HTTP listener.
It is NOT supported in the new TOML configuration format.

Similarly to 'mod_cowboy', this code should be either removed or the
  missing TOML support should be added.

As there is no final decision yet (and to keep the steps small), the
  changes in this commit are reduced to minimum:
  the reload test is removed with the config files that are impossible
  to convert to TOML.