Fix naming clash
The cause of this 404: locally mkdocs serves this "configuration files"
page as localhost:8000/Advanced-configuration/, while on github.io we
have .github.io/MongooseIM/advanced-configuration/configuration-files/,
mind the last part of the link. The latter way of handling it is the
correct one.

If we want to keep this chapter as “configuration files”, in mkdocs.yml it should be:
  - 'Configuration':
      - 'Configuration files': 'advanced-configuration/Configuration-files.md'
andalso  mv doc/Advanced-configuration.md doc/advanced-configuration/Configuration-files.md.
The last step is to update its link in Jingle-SIP-setup.md and everything is clean.