ejabberd_sm:get_session/ to return session records
Up to now, get_sessions/0,1 returned `[ses_tuple()]`, but
get_sessions/2,3 returned `[session()]`, which to begin with seems
inconsistent, and perhaps that is the biggest benefit of this change.

Profiling data from the mongooseim_one_to_one scenario didn't really
show any difference, sometimes increases of 0.01%, sometimes decreases
of 0.01%. This is sadly expected, most often that scenario doesn't cover
the paths where this is used.