[src] Query extensions in base mod_inbox_rdbms
There's a bunch of hardcoded calls to mod_inbox_bkpr, for selecting more
fields, filtering from more patterns, and returning more values, that I
tried to keep out of the base mod_inbox_rdbms, that might suffer
modifications in open source, keeping all bkpr logic just in
mod_inbox_bkpr.

The most challenging part is the tuple returned from the sql driver,
there's no way to pattern-match a variable-length tuple and tell the
extensions to parse all the suffix, I had to change the tuple size and
hardcode the extended fields.

I considered doing a tuple_to_list and pattern-match on that, but it
would be a performance overkill, I don't think we need to go that far.

Note that, as only the extension knows how the fields were requested,
only the extension knows how to process them and in which order.