[tests] Implement all possible extension tests
This suite just has every single scenario covered by the extension, it
is very exhaustive and I tried to write the code in an organised and
predictable way, full of comments describing the intention, spec
annotations, and what not. I'm not reflecting here how did this came to
happen, just the final state of the test-suite.