Use prepared queries in mod_event_pusher_push
Also: split the 'disable' callback

There are two distinct use cases now:
  - disable/1 for cleanup on user removal
  - disable/3 for disabling pubsub jid with an optional node

Before it was possible to specify the node without the pubsub JID.
However, this was unused and there was no use case or test for it.
XEP-0357 also does not allow this use case.
There was no point in introducing an unused prepared query.

Other changes:
  - Minor refactoring
  - Fix type specs