Separate backend types from front-end types
If we want to support more backends, the interface between them must be
pretty erlang idiomatic, so keep it integers and booleans on their
interface.

Also, note that the start function was using the wrong key number for
backend, it's the first of a tuple, not the second.

And one more thing, the accumulator contains the timestamp of the
incoming message, so we can skip calculating more timestamps along the
way.