changing format of hide_service_name option to local_config
hide_service_name configuration for specific host type (previously for specific domain) make no sense,
since we don't know the domain name at the moment when we fetch the value of the option. in fact (before
introduction of the host types) we were always using the value for ?MYNAME host, which was the first
item in the 'hosts' list.