Use a separate hook for MUC feature discovery
Motivation: instead of overloading one hook (disco_local_features)
  and running it both for the local domain and for the MUC domain,
  the logic can be split.
This eliminates the need to pattern-match on domains for all handlers.
The code is simpler and it is harder to make a mistake,
  e.g. by forgetting to pattern-match.
This way MUC features would never leak to local features and vice versa.

NOTE: This commit fixes the tests, they should pass from now on.