Disallow queries for another user's roster
According to RFC6121, section 2.3.3, it is forbidden to send roster set
to another user's JID.
In the current implementation there was only a check if the domain exists,
which would leak information about other tenants served by the system.

Handling of roster get to another user's JID is unspecified by the RFC,
but it makes most sense to handle it the same way - it is better
to return an error than to accept such stanzas, which are most likely
sent by accident.