adding some TODO comments
The @~2 commit is cherry-picked from beekeeper's repo. it addes declaration of dynamic_domains feature support to some modules. To see the full list of the changed modules use this command:
   git show --name-only  --pretty="format:" @~2 | sed -r "s_.*/(.*)\.erl_\1_p;d" | sort

The conversion of 'mod_commands' and 'mod_muc_light_commands' modules seems to be done, however there's no integration testing enabled for that modules in dynamic_domains.spec

Other modules (except mod_mam_meta) are dynamically initialised during the execution of the tests currently configured in dynamic_domains.spec, however there's no waranty that all of them are fully converted and tested. So ideally we should recheck all of them, here is the list of that modules:
 * mod_mam
 * mod_mam_cache_user
 * mod_mam_meta
 * mod_mam_mnesia_prefs
 * mod_mam_muc
 * mod_mam_muc_cache_user
 * mod_mam_muc_rdbms_arch
 * mod_mam_muc_rdbms_async_pool_writer
 * mod_mam_rdbms_arch
 * mod_mam_rdbms_async_pool_writer
 * mod_mam_rdbms_prefs
 * mod_mam_rdbms_user
 * mod_muc
 * mod_muc_light
 * mod_muc_log