Create a configurable FIFO cache
This uses the strategy of keeping a set of ets tables that are rotated
periodically. This means that TTLs have a very coarse-grained precision,
but book-keeping is considerably simpler.

Many things can be done from here, like whether we want to create a
`does_user_exist` hook where this and ejabberd_auth hook and inbox and
mam query, or if we want to add to the cache the calls from
ejabberd_auth and mam and inbox query here directly, or through a hook,
etcetera.