Reorder remove_domain and remove_user hook order
As described in a comment in PR3163:

We must ensure that remove_user and remove_domain hooks are executed
first for ejabberd_auth, after that for mod_cache_users and only then
for all other modules. So there is no chance that we asynchronously
re-insert the removed user in the cache, or insert any permanent data
(e.g. MAM, Inbox) for the removed but still cached user.