allowing only external function references in gen_hook interfaces
for OTP versions prior to 23, local function references in format like fun some_fun/2 behave in the same way as anonymous functions. different places in code generate different references. in addition to that local fun references are not code-reload safe, so it’s better to use external fun references only.

It's still a good idea to use function references, because it allows dialyzer to check the handler function specs.