Reset last_event_id once mongoose_domain_core starts
This tells to the service, that it needs to read the complete database.
But it assumes that the service process is down, so it would not
overwrite the last_event_id we set here.

To ensure that the service is down if the core is down,
the core and service processes should be supervised with rest_for_one strategy.

Which is in a separate PR.