Test inbox system messages under diverse domains
For this, we need alice_bis. In the default config, alice_bis is
configured under a separate host_type, which needs to be initialised;
in the dynamic_domains config, it is the same. For both cases, we need
to make sure alice_bis won't get aff_changes stored.

In this test, localhost.bis is configured as not storing affiliation
changes. So, when alice_bis gets an affiliation change, if it is not
correctly detected as a system message, it will be stored as a normal
message, never querying the correct configuration for aff_changes. If it
is correctly detected as a system message though, the aff_changes
configuration will then make it not be stored.