Add interface to accumulate affiliations
Very often, the affiliations of a room will need to be checked many
times. Imagine a room of a thousand people, and one sending a message:
if any handler on user_send_packet needs to know the affiliations, they
would be then queried again when the router triggers the process_packet
callback for muclight.

Such a handler happens often when we're checking if we can access the
room, like in the case of smart_markers: then we could be introducing a
race condition, when a user can react in a certain way in his
user_send_packet hook, but then muclight finds out the user has just
been kicked out.

This can also be used as an optimisation: inbox is making a DB query for
each user when muc_light routes the package to them, but we could be
making a single DB call in the context of the sender with all the
receivers queries at once, hence reducing round-trips considerably.