Handler records don't need the key
Handlers are all queried by a known key, and when the ets is fetched
by the key, the handlers are all deeply-copied, no sharing of keys,
which means the hook key, with all the host types, are copied into
the reader many times. This field can be entirely dropped,
it is not used by run_fold.