Store ACL specs in config
Motivation: rules and other options are stored in the config.
The ACL table required special handling in many places
and it will be easier to move to persistent terms if all options
are stored in the config - the code transformation will be more systematic.

Functions for modifying the ACL specs are removed
because they should not be modified in a running system.