Take the accumulator in the room hooks
This allows to have the accumulator in hooks like is_muc_room_owner and
can_access_room, so that if affiliations where already fetched, they
will be present in the accumulator, and we then move forwards in
preserving the consistency of a message processing pipeline: we want to
avoid the case when, during this pipeline, a subsystem fetches the
affiliations, take some action, and before the next subsystem acts, a
concurrent process modifies the affiliations: then the following
subsystem may find out that the user has been removed from the room and
reject the request, therefore getting into an inconsistent state.

This also helps preventing more duplicate, and usually idempotent, DB
queries.