Make the affs not permanent in the acc
On broadcast to connected users, there's a chance to build a copy of a
potentially long list of affiliations, to potentially many connected
users, which would make broadcast for this c2s process too slow.
Risk of an inconsistency for the recipients is too small to be worth the
trouble, they might as well fetch their own affiliations.