Use stored name instead of regenerating it on every use
I suspect this might also have somehow better concurrency
characteristics, as generating the name generates atoms, and therefore
needs to grab locks on the atom table, which is not optimised for
writes, even if the write is idempotent as the atom already exists, but
this one is mostly guessing. Sequential performance is surely much
better.