Create batch_worker abstraction
This is a gen_server, whose purpose is to be part of a worker pool. The
server accumulates tasks to take action to, until the number of tasks
reaches a configured threshold, or, a timeout is fired since the first
task that was requested in this batch (this is to ensure that no queue
awaits too long for action).

Note that we don't trigger timeouts unless the queue is not empty, to
avoid waking up an unused worker.

The worker also takes a callback, and a map with any metadata the
callback may need. When is time to flush a queue, it is this callback
that will be ran, taking the queue in the order in which it has been
submitted, and the optionally provided metadata.

Most of the code for this worker is learnt from MAM's async workers,
with some changes or learnt lessons:
* Note that this worker sets info logs on start and terminate.
* There is also a debug log on queue flushes.
* The terminate callback flushes the task queue if there's any task
  pending, to try to avoid losing tasks inadvertently.
* The format_status callback prevents printing the actual tasks on
  crashes or system state lookups, as the tasks might contain data
  private to clients under any GDPR-like legal constrains.
* Garbage collection is triggered after every flush, first to ensure
  saving memory, but second, to ensure the release of any binaries this
  worker might be holding.